+++ data-raw/ald_demo.R
@@ -13,4 +13,12 @@ ald_demo <- left_join(
 
 ald_demo$year <- as.integer(ald_demo$year)
 
+ald_demo <- ald_demo %>%
+  group_by(name_company) %>%
+  mutate(
+    id_company = as.character(cur_group_id()),
+    .before = 1
+  ) %>%
+  ungroup()
+
 usethis::use_data(ald_demo, overwrite = TRUE)